moved dataGenerators to chartUtils
