+++ oya-web/src/report/PieReportTotal.tsx
@@ -2,31 +2,10 @@ import { Box } from '@mui/material';
 import {
   Tooltip, ResponsiveContainer, PieChart, Pie, PieLabelRenderProps,
 } from 'recharts';
-import { ActivityTotalReport } from './reportUtils';
 import { Activity } from '../apiService';
 import RechartsTooltip from './RechartsTooltip';
 import { useReportContext } from './ReportProvider';
-
-type DataRe = {
-  name: string;
-  activity?: Activity;
-  value: number;
-};
-
-export const generateDataRe = (atrm: ActivityTotalReport, activities: Activity[]) => {
-  const data: DataRe[] = [];
-  activities.forEach(a => {
-    if (a.parentIds.length !== 0) return;
-    const r = atrm[a.id];
-    if (r && r.time)
-      data.push({
-        activity: a,
-        name: a.name,
-        value: r.avgPerAllDays,
-      });
-  });
-  return data.sort((a, b) => b.value - a.value);
-};
+import { generateDataRe } from './chartUtils';
 
 const RADIAN = Math.PI / 180;
 const renderCustomizedLabel = (props: PieLabelRenderProps) => {
